Use short month for FormattedDate
Because now server incorrectly renders short date when long is defined
- https://github.com/andyearnshaw/Intl.js/issues/179

Will be fixed soon, but people can be confused with “React attempted to
reuse markup in a container but the checksum was invalid.”

+++ src/browser/intl/IntlPage.react.js
@@ -54,7 +54,7 @@ class IntlPage extends Component {
         <p>
           <FormattedDate
             value={Date.now()}
-            {...{ day: 'numeric', month: 'long', year: 'numeric' }}
+            {...{ day: 'numeric', month: 'short', year: 'numeric' }}
           />
         </p>
         <p>
